Move effects to separate files
