+++ src/effects/disco.rs
@@ -1,0 +1,21 @@
+use std::{sync::atomic::Ordering, thread, time::Duration};
+
+use rand::Rng;
+
+use super::EffectPlayer;
+
+pub(super) struct Disco;
+
+impl EffectPlayer for Disco {
+	fn play(manager: &mut super::EffectManager, _direction: crate::enums::Direction, _rgb_array: &[u8; 12], _speed: u8, _brightness: u8, thread_rng: &mut rand::rngs::ThreadRng) {
+		while !manager.stop_signals.manager_stop_signal.load(Ordering::SeqCst) {
+			let colors = [[255, 0, 0], [255, 255, 0], [0, 255, 0], [0, 255, 255], [0, 0, 255], [255, 0, 255]];
+			let colors_index = thread_rng.gen_range(0..6);
+			let new_values = colors[colors_index];
+
+			let zone_index = thread_rng.gen_range(0..4);
+			manager.keyboard.set_zone_by_index(zone_index, new_values);
+			thread::sleep(Duration::from_millis(2000 / (u64::from(manager.keyboard.get_speed()) * 4)));
+		}
+	}
+}
